Move SDDPStats into dedicated file
